Move patch objects to the dedicated package
